adds css styling to searchpage and fixes functionality of search bar in SearchResultPage container
