+++ frontend/src/common/components/Footer/Footer.css
@@ -5,7 +5,7 @@
     text-align: center;
     left: 0;
     bottom: 0;
-    width: 100%;
+    width: 100vw;
     color: #333333;
     font-size: 1rem;
     font-weight: bold;
