+++ frontend/src/search/components/SearchBar/SearchBar.css
@@ -1,0 +1,35 @@
+.search-bar {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 100%;
+    max-width: 600px; 
+    margin-top: 20px;
+  }
+  
+  .search-bar input[type="text"] {
+    flex: 1; 
+    padding: 12px; 
+    font-size: 16px; 
+    border: 1px solid #ccc; 
+    border-radius: 4px; 
+  }
+  
+  .search-bar input[type="text"]:focus {
+    outline: none;
+    border-color: orange;
+  }
+  
+  .search-bar .search-button {
+    color: white; 
+    border: none;
+    padding: 13px 3px 13px 10px;
+    cursor: pointer;
+    border-radius: 4px; 
+  }
+  
+  .search-bar .search-button img {
+    width: 20px; 
+    height: 20px;
+    vertical-align: middle;
+  }
