+++ frontend/src/search/containers/SearchPage/SearchPage.css
@@ -1,3 +1,7 @@
-#search-page-container{
-    margin-top:200px;
-}
+#search-page-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: space-between;
+    min-height: 100vh; /* Ensure full viewport height */
+  }
