+++ frontend/src/search/containers/SearchResultPage/SearchResultPage.js
@@ -1,36 +1,38 @@
-import React, { useState } from 'react';
-import './SearchResultPage.css';
-import FilterOptions from '../../components/FilterOptions/FilterOptions';
-import FilterResult from '../../components/FilterResult/FilterResult';
-import Navbar from '../../../common/components/Navbar/Navbar';
-import SearchResults from '../../components/SearchResults/SearchResults';
-import Footer from '../../../common/components/Footer/Footer';
+import React, { useState, useEffect } from "react";
+import "./SearchResultPage.css";
+import SearchBar from "../../components/SearchBar/SearchBar";
+import SearchResults from "../../components/SearchResults/SearchResults";
+import recipeData from "../../../data/recipeData";
+import { useLocation } from "react-router-dom";
+import Navbar from "../../../common/components/Navbar/Navbar";
+import Footer from "../../../common/components/Footer/Footer";
 
 const SearchResultPage = () => {
-  const [showFilter, setShowFilter] = useState(false);
-  const [selectedFilters, setSelectedFilters] = useState([]);
+  const location = useLocation();
+  const queryParams = new URLSearchParams(location.search);
+  const query = queryParams.get("query") || "";
 
-  const toggleFilter = () => {
-    setShowFilter((prevFilter) => !prevFilter);
-  };
+  const [matchingRecipes, setMatchingRecipes] = useState([]);
 
-  const handleApplyFilters = (filters) => {
-    setSelectedFilters(filters);
-    setShowFilter(false);
+  useEffect(() => {
+    const filteredRecipes = recipeData.filter((recipe) =>
+      recipe.title.toLowerCase().includes(query.toLowerCase())
+    );
+    setMatchingRecipes(filteredRecipes);
+  }, [query]);
+
+  const handleSearch = (searchQuery) => {
+    const filteredRecipes = recipeData.filter((recipe) =>
+      recipe.title.toLowerCase().includes(searchQuery.toLowerCase())
+    );
+    setMatchingRecipes(filteredRecipes);
   };
 
   return (
     <div id="search-result-page-container">
       <Navbar />
-      <button onClick={toggleFilter}>Filter Search</button>
-      {showFilter && (
-        <FilterOptions
-          selectedFilters={selectedFilters}
-          applyFilters={handleApplyFilters}
-        />
-      )}
-      <FilterResult selectedFilters={selectedFilters} />
-      <SearchResults />
+      <SearchBar onSearch={handleSearch} initialQuery={query} />
+      <SearchResults matchingRecipes={matchingRecipes} />
       <Footer />
     </div>
   );
